+++ .prettierrc.json
@@ -1,0 +1,6 @@
+{
+   "tabWidth": 3,
+   "semi": true,
+   "singleQuote": false,
+   "bracketSpacing": true
+}
