+++ index.html
@@ -4,8 +4,8 @@
 <head>
 	<meta charset="UTF-8">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<link rel="stylesheet" href="css/null.css">
-	<link rel="stylesheet" href="css/style.css">
+	<link rel="stylesheet" href="./src/css/null.css">
+	<link rel="stylesheet" href="./src/css/style.css">
 	<title>Мемори</title>
 </head>
 
@@ -13,7 +13,7 @@
 	<div class="app">
 
 	</div>
-	<script type="module" src="./js/index.js"></script>
+	<script type="module" src="./src/js/index.js"></script>
 </body>
 
 </html>
