--- js/routes.js
@@ -1,2 +1,0 @@
-export const CHANGE_DIFFICULTY_PAGE = 'change_difficulty';
-export const GAME_PAGE = 'game';
