+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "format": "prettier --write ./src/"
   },
   "repository": {
     "type": "git",
@@ -15,5 +16,8 @@
   "bugs": {
     "url": "https://github.com/JuliaMatsievich/game-memory/issues"
   },
-  "homepage": "https://github.com/JuliaMatsievich/game-memory#readme"
+  "homepage": "https://github.com/JuliaMatsievich/game-memory#readme",
+  "devDependencies": {
+    "prettier": "^2.8.8"
+  }
 }
