+++ src/css/style.css
@@ -1,0 +1,128 @@
+@font-face {
+   font-family: "StratosSkyeng";
+   font-style: normal;
+   font-weight: 400;
+   font-display: swap;
+   src: url("../fonts/stratosskyengweb-regular.woff2") format("woff2");
+}
+
+.app {
+   background-color: #004980;
+   width: 100%;
+   height: 100vh;
+   display: flex;
+   justify-content: center;
+   align-items: center;
+}
+
+.popup {
+   width: 480px;
+   border-radius: 12px;
+   background: #c2f5ff;
+   padding: 50px 70px;
+}
+
+.popup__container {
+   display: flex;
+   flex-direction: column;
+   align-items: center;
+}
+
+.popup__title {
+   color: #004980;
+   text-align: center;
+   font-size: 40px;
+   line-height: 48px;
+   margin-bottom: 48px;
+}
+.popup__content {
+}
+.popup__content-difficulty {
+   margin-bottom: 67px;
+}
+.difficulty__container {
+   display: flex;
+   column-gap: 26px;
+}
+.difficulty__block {
+   color: #0080c1;
+   display: flex;
+   justify-content: center;
+   align-items: center;
+   width: 98px;
+   height: 98px;
+   border-radius: 12px;
+   background: #fff;
+   text-align: center;
+   font-size: 64px;
+   line-height: 72px;
+   cursor: pointer;
+   transition: transform 0.3s ease 0s;
+}
+.difficulty__block:hover {
+   transform: scale(1.1);
+}
+
+.difficulty__block input[type="radio"] {
+   display: none;
+}
+
+.difficulty__block input[type="radio"]:checked + label {
+   background: #7ac100;
+}
+
+.difficulty__block_checked {
+   color: #fff;
+   background: #7ac100;
+   transform: scale(1.1);
+}
+
+.difficulty__block label {
+   cursor: pointer;
+}
+
+.button {
+   display: flex;
+   width: 246px;
+   height: 48px;
+   justify-content: center;
+   color: #fff;
+   text-align: center;
+   font-size: 24px;
+   line-height: 32px;
+   border-radius: 12px;
+   background: #7ac100;
+   align-items: center;
+}
+.button:hover {
+   background: #5c9100;
+}
+.start__button {
+}
+
+@media (max-width: 510px) {
+   .popup {
+      width: 300px;
+      padding: 30px 50px;
+   }
+   .popup__title {
+      font-size: 30px;
+      line-height: 38px;
+      margin-bottom: 38px;
+   }
+   .difficulty__container {
+      column-gap: 16px;
+   }
+   .popup__content-difficulty {
+      margin-bottom: 47px;
+   }
+   .difficulty__block {
+      width: 78px;
+      height: 78px;
+      border-radius: 12px;
+      background: #fff;
+      text-align: center;
+      font-size: 44px;
+      line-height: 52px;
+   }
+}
