+++ src/js/routes.js
@@ -1,0 +1,2 @@
+export const CHANGE_DIFFICULTY_PAGE = "change_difficulty";
+export const GAME_PAGE = "game";
